prepare to NAND image buffer

+++ TestRunner/TestTaskRequestWrite.cpp
@@ -207,7 +207,7 @@ TEST_CASE("TaskRequestWrite: Handles Write Request Lifecycle")
     ImageMetadata metadata;
     metadata.camera_index = 10;
     metadata.timestamp = 0x12345678;
-    metadata.image_size = test_data.size();
+    metadata.image_size = static_cast<uint>(test_data.size());
     metadata.checksum = 0xff0000ff;
     mock_buffer.push_image(test_data, metadata);
 
@@ -311,7 +311,7 @@ TEST_CASE("TaskRequestWrite: Handles Write Request Lifecycle with Errors")
     ImageMetadata metadata;
     metadata.camera_index = 10;
     metadata.timestamp = 0x12345678;
-    metadata.image_size = test_data.size();
+    metadata.image_size = static_cast<uint32_t>(test_data.size());
     metadata.checksum = 0xff0000ff;
     mock_buffer.push_image(test_data, metadata);
